[CCAP-577] Move FileNameUtility to correct package
