+++ src/main/java/org/ilgcc/app/file_transfer/DocumentTransferRequestService.java
@@ -9,7 +9,6 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
@@ -20,7 +19,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.ilgcc.app.config.DocumentTransferConfiguration;
 import org.ilgcc.app.data.Transmission;
 import org.ilgcc.app.data.TransmissionRepositoryService;
-import org.ilgcc.app.utils.enums.FileNameUtility;
+import org.ilgcc.app.utils.FileNameUtility;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
